Some refactoring
A rough plan of attack:

(1) Publish this and make Whiley Compiler depend on it.

(2) Update Content.Ledger so it can be a real ledger supporting
transactions.  I think it can all be done using put() and putAll(),
where assigning a path to null is the same as deleting it.  One issue is
whether we want to support more complex forms of transaction, such as
copying, moving, etc.

(3) The Build interface disappears, but a repository would still
implement the ledger interface.  Potentially, we might pull Ledger out
as a top-level entity.  Potentially we don't even need this interface.

(4) DefaultRegistry Should become something like HashMapRegistry.

(5) Sink doesn't can drop remove() and add putAll() to given a
consistent interface for transactions.

(6) Potentially Sink.getContentRegistry() should be placed elsewhere as
well ?