Split out Build Graph #1
This splits out the build graph and related concepts, since they are not
coherent with the key responsibilities of this library.  That is, the
management of structured content across sources and sinks made up of
key/value pairs.

As such, the build graph concepts need to find their own home.